Update handleFromCypress function
When setting up Mirage + Cypress in this site (a gatsby site) we went through this, and the `headers.map` check failed. I don't know where we got `.map` from but it's not part of the [Headers API](https://developer.mozilla.org/en-US/docs/Web/API/Headers). This is the correct way to check the header.

Also add a note about duplicating passthroughs.